requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer==3.3.3
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-storage-blob==12.24.0
blinker==1.9.0
boto3==1.35.68
botocore==1.35.68
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
instaloader==4.13.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
Levenshtein==0.26.0
MarkupSafe==3.0.2
msrest==0.7.1
nltk==3.9.1
oauthlib==3.2.2
openai==1.52.2
openpyxl==3.1.5
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pyspellchecker==0.8.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
setuptools==75.2.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
